Create fetch function and utils file
