+++ examples/medplum-photon-integration/src/bots/sync-patient.ts
@@ -1,13 +1,8 @@
 import { BotEvent, getReferenceString, MedplumClient, normalizeErrorString, PatchOperation } from '@medplum/core';
-import { Address, AllergyIntolerance, ContactPoint, Identifier, MedicationRequest, Patient } from '@medplum/fhirtypes';
+import { AllergyIntolerance, Identifier, MedicationRequest, Patient } from '@medplum/fhirtypes';
 import fetch from 'node-fetch';
-import {
-  CreatePatientVariables,
-  PhotonAddress,
-  PhotonAllergenInput,
-  PhotonMedHistoryInput,
-  PhotonPatient,
-} from '../photon-types';
+import { CreatePatientVariables, PhotonAllergenInput, PhotonMedHistoryInput, PhotonPatient } from '../photon-types';
+import { formatAWSDate, formatPhotonAddress, getSexType, getTelecom, handlePhotonAuth } from './utils';
 
 export async function handler(medplum: MedplumClient, event: BotEvent<Patient>): Promise<PhotonPatient> {
   const patient = event.input;
@@ -92,12 +87,17 @@ export async function handler(medplum: MedplumClient, event: BotEvent<Patient>):
   }
 
   const body = JSON.stringify({ query, variables });
+  const result = await photonGraphqlFetch(body, PHOTON_AUTH_TOKEN);
+  await updatePatient(medplum, patient, result);
+  return result.data.createPatient;
+}
 
+async function photonGraphqlFetch(body: string, authToken: string): Promise<any> {
   try {
     const response = await fetch('https://api.neutron.health/graphql', {
       method: 'POST',
       headers: {
-        Authorization: 'Bearer ' + PHOTON_AUTH_TOKEN,
+        Authorization: 'Bearer ' + authToken,
         'Content-Type': 'application/json',
       },
       body,
@@ -108,44 +108,7 @@ export async function handler(medplum: MedplumClient, event: BotEvent<Patient>):
     }
 
     const result = await response.json();
-    if (result.errors?.[0]) {
-      throw new Error(result.errors[0].message);
-    }
-    await updatePatient(medplum, patient, result);
-    return result.data.createPatient;
-  } catch (err) {
-    throw new Error(normalizeErrorString(err));
-  }
-}
-
-async function handlePhotonAuth(clientId?: string, clientSecret?: string): Promise<string> {
-  if (!clientId || !clientSecret) {
-    throw new Error('Unable to authenticate. Invalid client ID or secret.');
-  }
-
-  const body = {
-    client_id: clientId,
-    client_secret: clientSecret,
-    audience: 'https://api.neutron.health',
-    grant_type: 'client_credentials',
-  };
-
-  try {
-    const response = await fetch('https://auth.neutron.health/oauth/token', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        Accept: 'application/json',
-      },
-      body: JSON.stringify(body),
-    });
-
-    if (!response.ok) {
-      throw new Error(`HTTP Error! Status: ${response.status} ${response.statusText}`);
-    }
-
-    const result = await response.json();
-    return result.access_token;
+    return result;
   } catch (err) {
     throw new Error(normalizeErrorString(err));
   }
@@ -246,31 +209,10 @@ async function getPhotonId(
     return undefined;
   }
   const body = formatRequestBody(resource.resourceType, rxcui);
-
-  try {
-    const response = await fetch('https://api.neutron.health/graphql', {
-      method: 'POST',
-      headers: {
-        Authorization: 'Bearer ' + authToken,
-        'Content-Type': 'application/json',
-      },
-      body,
-    });
-
-    if (!response.ok) {
-      throw new Error(`HTTP Error! Status: ${response.status} ${response.statusText}`);
-    }
-
-    const result = await response.json();
-    const id =
-      resource.resourceType === 'AllergyIntolerance'
-        ? result.data.allergens?.[0]?.id
-        : result.data.medications?.[0]?.id;
-
-    return id;
-  } catch (err) {
-    throw new Error(normalizeErrorString(err));
-  }
+  const result = await photonGraphqlFetch(body, authToken);
+  const id =
+    resource.resourceType === 'AllergyIntolerance' ? result.data.allergens?.[0]?.id : result.data.medications?.[0]?.id;
+  return id;
 }
 
 function formatRequestBody(resourceType: 'AllergyIntolerance' | 'MedicationRequest', rxcui: string): string {
@@ -306,60 +248,4 @@ function formatRequestBody(resourceType: 'AllergyIntolerance' | 'MedicationReque
     default:
       throw new Error('Invalid resource type');
   }
-}
-
-function formatAWSDate(date?: string): string {
-  if (!date) {
-    return '1970-01-01';
-  }
-
-  return new Date(date).toISOString().slice(0, 10);
-}
-
-function getSexType(sex?: Patient['gender']): 'MALE' | 'FEMALE' | 'UNKNOWN' {
-  if (sex === 'female') {
-    return 'FEMALE';
-  } else if (sex === 'male') {
-    return 'MALE';
-  } else {
-    return 'UNKNOWN';
-  }
-}
-
-export function getTelecom(system: ContactPoint['system'], person?: Patient): string | undefined {
-  if (!person) {
-    throw new Error('No patient provided');
-  }
-
-  const telecom = person.telecom?.find((comm) => comm.system === system);
-  const telecomValue = telecom?.value;
-
-  if (!telecomValue && system === 'phone') {
-    throw new Error('Patient phone number is required to sync to Photon Health');
-  }
-
-  if (!telecomValue) {
-    return undefined;
-  }
-
-  if (system === 'phone') {
-    return telecomValue.slice(0, 1) === '+1' ? telecomValue : '+1' + telecomValue;
-  } else {
-    return telecomValue;
-  }
-}
-
-function formatPhotonAddress(address?: Address): PhotonAddress | undefined {
-  if (!address) {
-    return undefined;
-  }
-
-  return {
-    street1: address.line?.[0] ?? '',
-    street2: address.line?.[1] ?? '',
-    city: address.city ?? '',
-    state: address.state ?? '',
-    country: address.country ?? '',
-    postalCode: address.postalCode ?? '',
-  };
 }
